refactor: change file naming strategy
